refactor: consolidate settings page UI
