--- src/features/settings/index.ts
@@ -1,2 +1,0 @@
-export { PreferencesForm, type PreferencesFormProps } from "./ui/components/PreferencesForm";
-export { usePreferencesFormState } from "./model/hooks/usePreferencesFormState";
