+++ src/pages/settings/model/usePreferencesFormState.spec.tsx
@@ -1,9 +1,3 @@
-/**
- * @file Verifies the "PreferencesForm with usePreferencesFormState" contract with automated examples.
- * The examples make the expected behavior concrete with cases such as "auto-submits boolean and
- * numeric field changes", "synchronizes dark mode when the preferences prop changes".
- */
-
 import type { Preferences } from "@/entities/preferences";
 
 import type React from "react";
@@ -13,24 +7,20 @@ import { render, fireEvent, screen, waitFor } from "@testing-library/react";
 import { expect, it, describe, vi } from "vitest";
 import "@testing-library/jest-dom/vitest";
 
-import { PreferencesForm } from "../../ui/components/PreferencesForm";
+import { SettingsView } from "../ui/SettingsView";
 import { usePreferencesFormState } from "./usePreferencesFormState";
 
-/**
- * Renders the test-only Preferences Form Harness component with controlled state or providers.
- * Individual tests reuse it to exercise realistic interactions without repeating setup code.
- */
-const PreferencesFormHarness: React.FC<{
+const SettingsViewHarness: React.FC<{
   preferences: Preferences;
   onSubmit: (preferences: Preferences) => void;
 }> = ({ preferences, onSubmit }) => {
   const preferencesForm = usePreferencesFormState({ preferences, onSubmit });
-  return <PreferencesForm {...preferencesForm} />;
+  return <SettingsView {...preferencesForm} />;
 };
 
 import { createPreferences } from "@/test/factories";
 
-describe("PreferencesForm with usePreferencesFormState", () => {
+describe("SettingsView with usePreferencesFormState", () => {
   const preferences = createPreferences({
     showHeader: false,
     showSwipeButtonList: false,
@@ -46,7 +36,7 @@ describe("PreferencesForm with usePreferencesFormState", () => {
 
   it("auto-submits boolean and numeric field changes", async () => {
     const onSubmit = vi.fn();
-    render(<PreferencesFormHarness preferences={preferences} onSubmit={onSubmit} />);
+    render(<SettingsViewHarness preferences={preferences} onSubmit={onSubmit} />);
 
     await userEvent.click(screen.getByRole("checkbox", { name: "Show header" }));
     await waitFor(() => {
@@ -79,12 +69,12 @@ describe("PreferencesForm with usePreferencesFormState", () => {
 
   it("synchronizes dark mode when the preferences prop changes", async () => {
     const onSubmit = vi.fn();
-    const { rerender } = render(<PreferencesFormHarness preferences={preferences} onSubmit={onSubmit} />);
+    const { rerender } = render(<SettingsViewHarness preferences={preferences} onSubmit={onSubmit} />);
     const darkModeInput = screen.getByRole("checkbox", { name: "Dark mode" });
     expect(darkModeInput).not.toBeChecked();
 
     const updatedPreferences = { ...preferences, appearance: { ...preferences.appearance, darkMode: true } };
-    rerender(<PreferencesFormHarness preferences={updatedPreferences} onSubmit={onSubmit} />);
+    rerender(<SettingsViewHarness preferences={updatedPreferences} onSubmit={onSubmit} />);
 
     await waitFor(() => {
       expect(darkModeInput).toBeChecked();
