+++ src/pages/settings/model/usePreferencesFormState.ts
@@ -1,37 +1,14 @@
-/**
- * @file Provides the settings feature's Use Preferences Form State React hook.
- * The hook combines state and operations behind one interface so components do not need to
- * coordinate services themselves.
- */
-
 import type { Preferences } from "@/entities/preferences";
 
 import * as React from "react";
 import { useForm, useWatch } from "react-hook-form";
-
-import type { PreferencesFormProps } from "../../ui/components/PreferencesForm";
 
 export interface UsePreferencesFormStateOptions {
   preferences: Preferences;
   onSubmit?: (preferences: Preferences) => void;
 }
 
-export interface UsePreferencesFormStateReturn {
-  preferences: Preferences;
-  fields: PreferencesFormProps["fields"];
-  maxNumberOfCardsToLearn: number;
-  cardInterval: number;
-}
-
-/**
- * Provides the preferences form state values and operations needed by React components.
- * Callers receive one focused interface without coordinating the settings feature's stores and
- * services themselves.
- */
-export const usePreferencesFormState = ({
-  preferences,
-  onSubmit,
-}: UsePreferencesFormStateOptions): UsePreferencesFormStateReturn => {
+export const usePreferencesFormState = ({ preferences, onSubmit }: UsePreferencesFormStateOptions) => {
   const { control, handleSubmit, register, setValue, subscribe } = useForm<Preferences>({
     defaultValues: preferences,
   });
@@ -49,7 +26,7 @@ export const usePreferencesFormState = ({
     setValue("appearance.darkMode", preferences.appearance.darkMode);
   }, [preferences.appearance.darkMode, setValue]);
 
-  const fields: PreferencesFormProps["fields"] = {
+  const fields = {
     showHeader: register("appearance.showHeader"),
     showSwipeButtonList: register("controls.showSwipeButtonList"),
     showSwipeFeedback: register("appearance.showSwipeFeedback"),
@@ -70,7 +47,6 @@ export const usePreferencesFormState = ({
   };
 
   return {
-    preferences,
     maxNumberOfCardsToLearn,
     cardInterval,
     fields,
