+++ src/pages/settings/ui/SettingsSection.tsx
@@ -1,9 +1,3 @@
-/**
- * @file Defines the settings feature's Settings Section presentation component.
- * The component renders props and reports user intent through callbacks while data access stays
- * outside the view.
- */
-
 import * as React from "react";
 
 export interface SettingsSectionProps {
@@ -13,11 +7,6 @@ export interface SettingsSectionProps {
   children: React.ReactNode;
 }
 
-/**
- * Renders the Settings Section user interface.
- * Groups related settings under an accessible heading with a title, description, icon, and child
- * controls.
- */
 export const SettingsSection: React.FC<SettingsSectionProps> = ({ title, description, icon, children }) => {
   const headingId = React.useId();
 
@@ -52,10 +41,6 @@ export interface SettingsRowProps {
   children: React.ReactNode;
 }
 
-/**
- * Renders the Settings Row user interface.
- * Pairs one setting control with its linked label and optional explanatory description.
- */
 export const SettingsRow: React.FC<SettingsRowProps> = ({ inputId, label, description, children }) => (
   <div className="flex min-h-touch items-center justify-between gap-4 px-4 py-3">
     <div className="min-w-0">
